requirements.txt
Django==1.8.3
djangorestframework==3.4.5
httpie==0.9.6
Pygments==2.1.3
requests==2.11.1
wheel==0.24.0